User-level infraestructure for platform-dependent sensing
Previously sensing was fully performed by the sensing kernel module.
This led to a somewhat hackish integration with preeexisitng sensor
drivers in the kernel (e.g. power, temperature).

This patch adds a sensing infrastructure to interface with preexisting
drivers without using our sensing module. It manages the sensing
windows (very similarly to what the kenel module does) so that
the sensing model for external sensors is the same as the one
provided by our kernel module. The power sensing capability is
still part of our kernel module but it will be removed and the
our kernel module will be responsible for only collecting perf.
data (so it can be more easily swaped with perf_tools in the future).

An implementation of the interface with the XU3 power sensor is
provided. The power sensing data is read using the new sensing
function (see basic.cc / InterfaceTest).