Removed reference wrappers from sensing interface
Let's say we have the code:

const Resource *r = ...
sense<WHATEVER>(r,wid);

We want the sense call to be equivalent to:

sense<WHATEVER,Resource>(r,wid)

so the overloaded version of sense that takes a const Resource* is
called. However, the second template parameter was being infered as
const* which made the initial call equivalent to:

sense<WHATEVER,const Resource*>(r,wid)

which called the overload version of sense that takes a reference (in
this case would be a reference to a pointer). This will cause either
a "ambiguous call error" or some other error since the second call
would be forwarding a pointer-to-a-pointer.

Anyway, now only the pointer verison exists and resources are passed as
pointer only.