Added combined ubench execution for predictor training
predictor_master.cc runs all training combinations in a single
application. Each combination is designed to run for about 100ms (this
can be updated at TARGET_UBENCH_RT_MS). To do that, predictor_master
should first run in calibration mode to find the number of iterations
for each ubenchmark that will produce that runtime.

For example, running predictor_master without args (calibration mode)
will produce something like this in the output:

calibration_args 16286538 1861616 3874209 2708202 3889304 2766697 5096724 2093288 4559645 8131556 27

This can then be used with predictor_master again to execute the
combinations with the given number of iterations:

bin_host_offline/ubenchmarks/predictor_master 16286538 1861616 3874209 2708202 3889304 2766697 5096724 2093288 4559645 8131556 27

Between the execution of each combination, the app sleeps for
TARGET_UBENCH_RT_MS. This gap should be used to identify each diffent
combination when looking at the execution trace. There should be no idle
time in the trace samples collected in the middle of the execution of
one combination.