Allows daemons to exit from within
Daemons can use the System::quit() to cleanly end the daemon execution
(instead of sending SIGQUIT from scripts/runtime/stop.sh)

script/runtime/wait_for_stop.sh was added to be used instead of stop.sh
to block until the daemon is expected to call Sytem::quit().

Note that these are additions and the previous daemon execution flow is
still available and is expected to be used in most cases.