Removing power of idle cores from tracing
When traing an app we want the 'power_w' column in the traces to have
the power of the core the application is running only. But multiple
cores might share the same power sensor, i.e., the same power domain, so
that measurement might include the power of other cores.

When tracing it is assumed that only the traced app is running so all
other cores should be idle. Given this assumption, we now do the
following:

- remove_idle_power.py uses previously collected idle power measurements
to remove from 'power_w' the idle power of the cores that are supposed
to be idle.
- trace_one-parse.sh was modified to always run remove_idle_power.py. If
the idle power profiling data is not available, it runs
get_idle_power.sh to obtain it