Fixed sensing module deadlock
A deadlock may happen with the vitamins_cpu_counters_acc_lock when the
sensing kthread is interrupted with one of those locks acquired. A
previous solution used try_lock to skip sampling when a deadlock could
happen.

Now a double-buffer is used to accumulate samples when context
switching. One of the buffers is always available during a context
switch so there won't be a deadlock or lost/skipped samples.