Avoiding hardcoded string for sensing info
Using a constant to refer to column names when dumping traces/info.
This is useful since we might want to parse these traces later and
we have a handle for expected columns names.

Also, to make the conversion between string names to/from sensing
data types, a fixed string name is associated with sensed data
types, along with helper functions.