Added parser for traces generated by TracingSystem
TraceParser (src/runtime/common/traceparser.h) provides a cleaned-up
version of the parse_csvs() function from src/offline_sim/inputparser.h

The main difference between TraceParser and parse_csvs() is that
TraceParse supports ONLY the formats generated by TracingSystem.

Other features provided (not present in the old parse_csvs()):

1) Traces are normalized.
The traces read using TraceParser are supposed to be used for model
training and offline simulation. In both case, we need to map samples
from one trace to another trace (of the same workload). This mapping is
done by looking at the number of instructions executed, so TraceParser
now adds minor ajustments to the number of instructions in each samples
so traces can be more easily matched. This is necessary since the same
exact workload may present variations in the number of instrucitons
executed depending on the core arch/frequency. This normalization
removes those. Note, this only fix very small variations; any unexpected
variation will trigger a warning.

2) Samples have a 'sense-like' interface to read its values, which can
be accessed though the TraceSensingInterace.

TraceSensingInterace provides an interface equivalent to
SensingInterface that pulls data from a trace instead of actual sensing
data.

The setTrace(...) functions should be used to set the trace that will be
used as source for the sense(...) functions. setTrace takes either a
single sample to extract data from, or a range of samples. For the
second case, all samples will be aggregated before they can be read by
sense.

So we know how to aggregate the data automatically. A field describing
the aggregation method was added to SensingTypeInfo. Additional helper
functions to get the sensing type string and aggregation method are
provided.