Bin-based predictor trainer work with new traces
Update the bin based predictor generator to work with the TraceParser
instead of inputparser.h definitions. Also removed the old use cases
since they won't be fixed to work with the new training code.

In addition to the necessary adaptations, these improvements were
made:

1) Changed method to detect traces' sample windows

The old approach was not very flexible and was missing some obvious
windows. The new method is more bruteforce-ish but does a better job.

obtain_windows(...) now take the min/max number of samples each window
should have. This can be used to emulate the previous
obtain_windows_single_sample()

2) Refactored bin-based pred. to use new interfaces

Using the new TraceSensingInterface to implementent the computation of
average samples and the BinFuncs. BinFuncs also moved to a separate file
for clarity.

3) A OO interface for bin-based prediction is provided

4) More flexible bins and funcs.

Refactored BinBasedPred::Bin into two classes: PredBin which is
independent from training data and TrainingBin, which specializes
PredBin with the list of sample used to train it (equivalent to previous
BinBasedPred::Bin).

Also removed the BinBasedPred::Metric class and replaced by a more
dynamic and automated BinFunc class.

5) Removed the old code that generated the C-base predictor

The code to perform actual runtime is in the bin_based_predict function which
was refactored to be more generic (good for both testing offline and runtime
predictions)

6) Saving predictor as a JSON file

7) Also ported the correlation app

Calculates the correlation between candidate metrics for prediction.
Helps idendify a good configuration of layers for the bin-based
predictor.

8) Added the StaticHWModel interface

This should be the final wrapper over lower level predictions.
In addition to wrapping the bin-based pred interface. It also
adds some small error correction for predictions.
Checks if previouly sensed values are similar. If that's the case, the
final predicted value will be a wighted sum of the value predicted by
the bin predictor and the previous sensed value. The weight depends on
how close the previous sensed sample is from the current one.