Added daemon to test predictors on odroid
Every epoch the daemon predicts the perf/power for the next epoch and
then calculates the error in the next epoch. Every three epochs we
switch between big/little cores.

In addition to the daemon, some helper scripts were also added to
train and test predictors. scripts/training/train_bin_based_pred.sh
runs the predictor_trainer app for the current configuration.
Other scripts are helpers to collect traces and test the predictor on
the odroid.