Replacing System by PolicyManger
System was a somewhat generic name. The System class  basicaly
instanciates and groups policies, providing the features to
invocate and manage such policies, so it was renamed to something
more meaniful: PolicyManager.

Additionaly, when registering a sensing window, a priority is also given.
The priority is used to define the execution order of windows that
have the same period. Windows with the same period and priority
are executed in the order they are created.

Other minor changes:

- Changed visibility of some fields for safety
- Refactored common code in PolicyManager constructor
- Moved window manager to the framework folder. The SensingWindowManager
  is tightly coupled with the PolicyManager so it makes more sense to have
  it in the framework folder.