Defined Policies and Models
Defined the generic Model and Policies classes. A Model models resource
management policies that issues actuation decisions. A Policy is the same
as a model (Policy is also a subclass of Model) the only difference is that
actuations issued by a policy affects are performed on the system, while
actuations issued by a model affect only performance/power predictions
(to be made by the future senseIf interface).

To use these classes, the user need to create a subclass of Policy/Model
and define an execute() to implement the model/policy. Both models and
policies need to define a period and a priority that will be used to
decide when they execute. The PolicyManeger nows provides the functions to
register policies and models (the functionality to automaticaly execute
models will be added in a future patch that implements the senseIf
interface).

The execution of policies can happen together with the execution of other
sensing window handlers registers separatelly using
windowManager()->addSensingWindowHandler. The order is again defined by
the period and the priority.

Additional minor changes:
 - WindowManager uses multiset instead of ordered linked list