Added definitions for senseIf
senseIf returns predicted sensed data given actuations issued by
tryActuate. sense also may return predicted data if it's called when
running a policy as a model. The sense implementation now is a wrapper
to Impl::sense and senseIf that checks the current context to decide
which one to call. Implementers should now define specializations for
SensingInterface::Impl::sense*. An additional version of sense is added
that does not require passing the wid (it automatically uses the wid of
the current window). senseIf by definition only works with the wid of the
current window.

senseIf() however is not implemented yet.