Implemented tryActuate and senseIf
The basic idea is that policies use sense() to fetch sensed data, and
senseIf() predict how the sensed data will look like in the future given
an actuation decision. To inform about a possible actuation decision
without actually performing the actuation, the tryActuate() function is
used.

The connection between tryActuate() and senseIf() is done by the
ReflectiveEngine class. tryActuate() informs the ReflectiveEngine about
the requested actuation and the ReflectiveEngine internally stores it and
also keeps track of all resources that could have been affected by the
actuation. When senseIf() is called, it checks if the resouce we are
sensing was affected by a tryActuate(). If not, then it just returns the
same as sense(), otherwise it calls ReflectiveEngine::predict* functions
to predict the effect of the actuation on the sensed data.

With the senseIf()/tryActuate(), we properly defined the role of the
BaselineModel (previously StaticHWModel) and made it a template class that
takes different HW and Sharing models. The role of the baseline model is
to provide metrics that can be later used to estimate values returned by
senseIf(). As of now, an implementation of the baseline model is expected
to return per-task IPC, Power, and TLC, which can be combined to model the
performance/power of other resources (see reflective_pred.cc).

The BaselineModel model takes as input the current sensed data, current
actuation values, and future actuation values. Future actuation values
are both taken from tryActuate() calls issued by the current policy as
well as from possible actuations that can be made by other policies that
execute in the middle of the sensing window of the current policy.

To figure out future actuations we run these "finer-grained" policies
(they have an execution period smaller then the current policy) like
they are models (i.e. these actuations won't affect the system: actuate()
behaves like tryActuate()). Note the how ReflectiveEngine::Context
is set and used by the actuation and sensing interfaces to know if the
policy is actually being executed or just emulates as a model in the
scope of another policy.

In summary, the senseIf() flow is as follows:

 1)Policy models predict which actuations are going to be performed in
   in the future. If a model is undefined, we just assume the current
   values provided by actuate/tryActuate will not be changes during the
   future window and those are used. Notice that, for actuations that
   change during the window, the predicted value will often the "average"
   value if possible, or the most frequent value
   (define by ActuationTypeInfo::AggregatorType).

 2) predict() functions take the actuations predicted by step (1) and
    feed them to the BaselineModel. The metrics returned by the baseline
    model are then used to predict the requested metric returned to
    senseIf()

Addtionally the following minor changes were performed to allow the
implementation of the ReflectiveEngine:

- Refactored SensingInterface::Impl int a different header to break the
  dependency cycle:
     SensingInterface -> Models -> SensingInterface
  Now it is:
     SensingInterface -> Models -> SensingInterfaceImpl
                      -> SensingInterfaceImpl
  Same for the ActuationInterface.

- Modified policy_test daemon to test basic reflection using the dummy
  sensors/actuators

- Added example of reflective DVFS

- Added model_test to test errors of senseIf