Changing PolicyManager initialization process
Changes the way PolicyManager is created to be more friendly to offline
simulations.

Order and scope functions were called before:
    PolicyManager()
        init sys_info
        SensingWindowManager()
            SensingModule()

How it is now:
    SensingModule()
        init sys info
    PolicyManager(module_ptr)
        SensingWindowManager(module_ptr)

Since the sensing module and sys_info are all created beforehand and then
passed to SensingModule, there is no need to have #ifdefs inside
PolicyManager to handle the OfflineSensingModule.

Also the PolicyManager and SensingModule are not singletons anymore
(though the LinuxSensingModule prevent more then one module to exist), so
it's possible to have multiple PolicyManager instances in a offline
simulation. Access to PerformanceData from static functions (e.g. the
sense() functions in performance_data.cc) is done through thread_local
pointers initialized when new threads call the
SensingModule::resgisterAsDaemonProc() function.

deamonizer.h was updated to properly create the sensing module for the
Linux daemons.