Partial implementation of new offline sim
Making the offline sim work without the removed core_legacy would require
too much rework, so it was removed.

A new implementation that works better with the current implementation of
the PolicyManager and the new trace parser was added as replacement. The
current state of the new implementation allows policies to run and
supports the task mapping and DVFS actuations. Currently, summing-up the
sensed data per window is not supported (there are some placeholders/TODOs
in the code for that).

Other modifications:

- src/runtime/common/semaphore.h
    by default semaphore is locked

- src/runtime/common/traceparser.h/.cc
    added helpers to check if traces have a specific perfcnt

- src/runtime/framework/actuation_interface.h
- src/runtime/interfaces/linux/actuation_interface.cc
- src/runtime/framework/policy.cc:
    ActuationInterface::construct takes a ptr to the sensing module
    instead of sys_info (which can be fetched from the sensing module)

- src/runtime/interfaces/common/sense_data_shared.h
    Fixed typo

Next steps:

1) Added the code to sum-up sensing window data. Mostly copy the code from
the older OfflineSensingModule implementation with the adaptations to get
data from the new traces.

2) Integrated back with LinSched to support multiple tasks per-core.