Multiple Inputs Multiple Outputs example
This  commit includes all necessary files to replicate the design
process of a 2x2 MIMO controller in Matlab. example training files is
included with the script ot read this data to Matlab compatible
matrices. A set of system identification result is included (ss7 is used
for the rest of the process). Estimator design script takes care of the
mimo design process. LQG controller simulink shows the simulation of the
designed controller for the state space plant extracted in system
identification phase.